requirements.txt
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.4.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
h11==0.16.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.3.8
pycparser==2.22
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PySocks==1.7.1
requests==2.32.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
